Fix noslow/A and add tps management
noslow/A was broken. After strenous testing I made some changes to
make it functional again. The hardcoded speed value is set to 0.300
but needs to be managed better. Let me explain why.

During my testing using the math formula in Paradox I found that the
players velocity capped out at 0.153 while sprinting on a flat platform.
The original speed value was set so low that had the code not been
broken it would have false flagged simply by walking forward.

So why am I setting the value to 0.300 instead of say 0.154? Couple
of reasons here. For starters, sprinting at a velocity of 0.153
and jumping would spike the velocity as high as 0.263 and in rare
cases slightly higher. The other issue is the order of execution
within the code. The script was averaging about 20 executions per
second so I ran into an issue with the controller animations not
switching states fast enough and the script flagging based on a
condition of tags because the tags were being read before the
transition of the anination controller.

I needed the animation controller to update before the flag check.
The reason for this is because while sprinting you don't violate
noSlow/A. The problem is when you jump while sprinting there is a
spike increase on the player's velocity which flags you for violating
noSlow/A.So knowing the capped velocity mentioned above this means
if you don't exceed 0.154 under conditions then you are good assuming
the hardcoded value was set to that value. Unfortunately, when you
jump, your velocity goes from 0.153 to as high as 0.263 as mentioned
above. This shouldn't flag the player do to the conditions that need
to be met with tags but it does because of the current issue with
operation of execution within the code.

TPS is the main culprit and although this patch corrects most of this
issue, it is still far from perfect. More work needs to be done to
optimize management here.

Performance has definitely been improved thanks to scheduling since
the amount of times we were executing has been reduced exponentially.