Update frostwalker
Instead of locking in a pair of leather boots, we detect what type
of enchanted boots they have and give them a new set of those boots
without enchantments and naturally lock in that item.

The frostwalker exploit usually uses enchanted diamond boots but
we go an extra step here. We replace with the same type of boots
because its amusing to see them figure out why their boots are not
enchanted anymore.