Added Anti-KB module!
This ones a bit different from other additions. Added a new command to the player using mark variant (you can take it and replace with query.is_sheared if you need to use a multi-data version), and added a new very useful tag (attacked), that stays around after a player for 0.6 seconds or 12 ticks. Final thing, in case you are worried that it will false flag if someone is trapped with blocks - It won't, because they still have a small amount of velocity.