Fix namespoof/B
This check didn't work. It is intended to kick players from the server if their username contains any unicodes or despawn them if it can't properly kick the player. Originally this returned with a TypeError: Not a function. This has been corrected and will now kick or despawn players with unicodes in their names. This includes Hosts as well.