Unlock items when setting for Frostwalker is disabled
Originally when you enabled Anti Frostwalker it would detect the boots, then replace them, and lock them in your item slot to prevent an exploit. However, if you disabled the setting later on then the boots were still locked to your slot and it couldn't be removed so we check now if you have locked boots and remove the locked item to allow the player to play freely again if the setting is disabled.