When disabling anti frostwalker target everyone
Using target selector @s wont work because this will only target the player executing the command to disable anti frostwalker. We need to execute it for everyone when disabling. I wish there was a better way to do this. Hopefullly the upcoming beta gametest features will make this far more superior and this method can be abandoned then.