[HOTFIX] Fixed a bug that broke Paradox
Scoreboard objectives and score values were not being added to the player upon joining the world. Timing is increased to give more time to add them to the player as testing showed that they were being executed too quickly while the player was joining and consequently not being applied to that player. Another issue was with checking if gametest were enabled or not. Due to some changes with 1.18.10 in the Gametest Framework World.getPlayers() seemed to return as an iterator and not an array. This resulted in complete failure while executing gametestcheck.js by causing the script to fail in performance. With this patch, Paradox should be officially compatible with 1.18.10 now.