Target players in all dimensions
It seems to be pretty common in the Minecraft community to target players only in the Overworld dimension. For some reason we don't bother with the Nether or The End dimensions. Gametest offers the ability to determine what dimension the player is in simply by using the player class and dimension type. So now we can target players whether they are in the Overworld, Nether. or The End.