Update player and npc entities
Making NPC not able to be summoned is futile. We have a toggle for Anti NPC that is disabled by default but can be enabled at the users discretion. Furthermore all this is doing is preventing the use of the /summon command for this entity and causing Minecraft to report a syntax error but you can still spawn them by giving yourself a spawn_egg item. Player entity was simply diffed with the latest 1.18.10 file and merged.