Delete entities from Paradox
I do not see any reason for these to be here in my opinion. They essentially try to prevent spawning/summoning, but they can still be spawned. If we don't want to spawn these in then I think it should be left at the discretion of the user and should be handled how we handle NPC with Anti NPC. Possibly future feature.