Changes to chatfilter
With Gametest we don't need to use tellraw to filter chat messages. So instead of using tellraw we can use the property message: string. Unfortunately I could not find a property to modify the players name without using tellraw. Tests showed that player.nameTag returns the same as msg.sender.nameTag but there doesn't seem to be a specific property of some kind that allows msg.sender to be changed. So in cases where players change player.nameTag we use tellraw instead.