Add Xray to Gametest
When a player finds and breaks ancient debris or diamonds then it will notify staff if cheat notifications are enabled. It will tell the staff which player, what block type, and their current location so staff can watch them, should they choose to do so, to see if they are hacking to find the materials.