Cleaned up the help menu
Due to 569262aaa848126c89a4a1194f84397f579ab006 Paradox now officially depends on Gametest. Information that the help menu produced related to not having the framework enabled is now outdated and no longer supported. I have set it up to where it tags the players with gametestcheckapi and the help menu will use this tag to let the user know if they do not have the framework enabled.