Fix scoreboard creation when logging on
For unknown reasons to me the list needed to get split up. Not all entries would complete in the original compiled list. However, this fixes it. I plan to revisit this and maybe find another solution that I feel is cleaner. Main thing is getting Paradox up and running again for more development.