Fix error with FlyB
The segment added to illegalitems_a.js causes an error since it looks for a target that matches the conditions and can't find any. To solve that issue it is placed inside a try/catch method. Also moved it back to FlyB to prevent miscellaneous code. In doing so I created a new function named time then subscribed that function to a tick event. Although if it doesn't matter if it runs every 2 seconds then it would be better to remove this and place it inside the FlyB function instead since this will currently execute 20 timers per second as is.