Update list of illegal items
The list here focuses a lot on items only obtainable in creative mode or literally not obtainable at all unless hacking or using commands. This may not be a fully compiled list but it does cover the more common basics. The items removed either seemed not important or was part of the game experience and would break that experience if included. Given the nature of this list users are able to easily add or remove items as they see fit.