Major overhaul and port of the freeze feature to Gametest
This will tp a player above the clouds in the overworld and place barriers around them so they cannot leave. For additional security it will check to see if they moved from that location by cheating or hacking and will force them back inside the barriers. If they log off and rejoin it will persist to keep them at that location. They will not be released unless done so by the user with permission to use Paradox.

Regardless of which dimension the player may be in or where they are located within the world it will record their current coordinates and dimension. Then it will move them to the overworld where they will be imprisoned. When released, it will send them back to the coordinates and dimension they were originally located at prior to being frozen. The recording of their original coords and dimension is to insure we handle their positions more efficiently. Especially if and when they decide to log off and rejoin. We still have that information and can use it accordingly.

This feature officially requires the Gametest Framework.