Change Reach/A and add Reach/B
The original reach/A would check if a player attacked another player within a 5 block radius. Not a fan of the method used for this feature so I got rid of it for the time being. The new reach/A now checks if a player places a block more than 7 blocks away from their current position and will remove that block if they do.

Reach/B will check if a player breaks a block more than 7 blocks away from their current position. The broken block will be restored if flagged.