Removed killaura/b
I’ll be honest, it false flagged for me a few times when I was attacking and dying at the same time, but I assumed it was rare. After more testing, however, it flags roughly 70% of the time when attacking and being killed, which is enough for me to remove it completely. I plan to add a better version soon, but I wanted to get this fix out quickly.