Bugfix for ban
When banning a player using the command !ban it called for the function banMessage which is imported from utils.js. banMessage will simply kick the player and is supposed to be used in conjunction with the player having the tag for isBanned. In this particular instance the tag for isBanned wasn't being utilized as it should have. We don't need to directly call on that function here since Paradox automatically checks players for this tag and will call for the function on its own. So instead, tag them and let Paradox handle the rest from here.