Include reason when auto banned by nuker/a, illegalitems/A, and scaffold/A
The following three features will automatically ban a player when flagged. Before it wouldn't say by who or for what reason which may bring confusion as to what happened, why it happened, and who did it. So let the player know that it was Paradox who kicked them out and why Paradox did it depending on the type of flag that is triggered