Add antispam
This has a 2 second cooldown between chatting. If you chat multiple times within that cooldown period then Paradox will notify you to slow down. If a player starts spamming the same thing repeatedly then Paradox will just ban the player or kick as a fallback if it detects a match.

This doesn't stop a bypass used by hackers to spam. This is where the 2 second cooldown comes into place. If the spam is outside this 2 second cooldown window then there is nothing Paradox will do. It will be up to staff to intervene and kick the player.

You can disable antispam if you do not wish to use it by disabling it in config.js. You can also modify the cooldown period between messages from players in config.js as well. The value for the cooldown is based on 20 ticks per second. So 20 would be 1 second, 40 would be 2 seconds, 60 would be 3 seconds, 80 would be 4 seconds, etc. Set the value to what will cater to you. By default it is set to 40 (2 seconds).