Temporarily disable ban in nuker_a.js
For now, just flag the player instead of booting them since Gametest is making it slightly difficult to overcome potential false flags with this event. The false flags appear to be rare so at the least staff can TP to that players location and watch them. If visual confirmation is made, then take action.
Hopefully I can work this out at a later time.