Refactor Fly/A and delete Fly/B
Fly/B is basically Fly/A now. The code is updated to record the players position and store it. As they fly around for 6 seconds or more it will teleport the player back to their original location and keep teleporting them there until they land back on the ground again. This checks for only players who are in Survival mode. If they are in Creative or Adventure mode then it will ignore them.