Update manifest.json
This will be the beginning of including the version of the pack in the name of the pack so when installed users can clearly see which version they have without any confusion. This will be good given the situation that each release installs itself as a separate pack since the versions between each one differ from one another.