Update Jesus/A and delete Jesus/B
The original code was not sufficient at all. This change better detects those who are running across water or lava blocks then flags them. If you are riding in a boat, flying with Paradox permission, Vanished with Paradox Permission, or simply Paradox Opped,  then it will not trigger.

Key notes regarding this change:

In a vanilla world Mojang allows a player to essentially bunny hop across water without actually swimming. In a real world scenario this defies logic. So this change will flag a player if they bunny hop because a known hack by Horion to walk across water or lava is to bunny hop like an autoclicker. So if you want to move across the top of the water like you would in the real world then you need to make a boat or you will be flagged. This will not ban a player. That decision will be left at the discretion of the staff using Paradox in their world.