Update values for reach in config.js
When playing on mobile devices the reach extends out much further than on PC and consoles. There is no way of knowing which platform they are on that I am currently aware of so we have to give some lead way with PC and Console users to prevent undesirable flags on Mobile devices.