Revert values for reach/a and /b and do not flag
They were originally set to 7 which was fine for platforms
like PC's but this caused undesirable results on other
platforms such as mobile since they have a much further
reach. Raising the value for mobile consequently caused
these features to be completely useless on PC.

The best approach would be to determine which platform the
user is on then apply the correct values based on that return
to accomodate all platforms more efficiently and effectively.

Unfortunately there are no concrete ways to detect the platform
from a behavior pack unless Mojang makes it possible. So
this currently leaves three options at the moment. Option
one would be to throw out the features entirely, option 2
would be to use a bot to determine their platform, and option
3 would be to target the PC platform, and force mobile users
to be resticted to the same degree as on PC.

Option 1 is plausible but looking to avoid this. Option 2
is outside the scope of this project and is something a user
would need to implement for their own personal gain. Option
3 seems like the logical choice here and it doesn't hinder
gameplay on Mobile.

Going with option 3 we have to remove the flags to prevent
spamming of false flags from mobile users. So as of now, users
who do trigger A or B will just see either the block disappear
or be restored depending on the event kicking off and no
warning will be given either to the user or staff since this
particular instance doesn't necessarily warrant one anyways.