Make modules in config.js accessible while in the world
This commit brings forth the ability to enable/disable the modules found in config.js. The command for !modules will not include these in its list since these changes are specific to gametest only. Once "modules" is brought over to gametest then the changes made here can be included to allow users to actually see what is on and what is off.

Major note to take away from this change:

These particular modules are preset to true by default inside config.js. This means they are enabled automatically when a new world is created or if you reboot a server. So if you turn one off it will not persist after a server reset. You will have to disable that feature again. With that said, if there are any features you know you do not wish to be turned on at any given time and would like for that to persist after a server reboot then it is recommended to modify the file, config.js, and set them to false to get best results for your catering. This commit is a WIP for the future to come. I anticipate that these changes will open the door for new opportunities down the road.