Add new command !banwindow
This feature is turned off by default. This feature will allow staff to turn it on and disable the server ban in Paradox. This is to give the staff the opportunity to let server banned players to rejoin the world so they can unban them. This does not include global bans. Don't forget to turn it back off again.