Fix bug when clearing old ban tags
Tags were not properly being cleared and therefore would error if the tag already existed. This consequently caused try/catch to fail during try and fall back to catch where it would just despawn the player. This patch corrects this issue to insure tags are properly cleared if they exist to prevent the unintentional error.