Fix illegalitems/B
With the changes to the chat filter it broke this feature and would error when attempting to clear the item from the player. All because of double quotes lol.

Additionally, I added another condition to make sure that only Player entities execute the code. If it is a different entity then it will abort.