Fix tag ranks
Where to begin with this...

Let's just say the code was flawed because I am only human.

Went ahead and added additional code so the tag rank will refresh when changed or joining the world. Prior to this it wouldn't update above the player's head unless they said something in chat.