Patch new hack with fish buckets
There is a new hack out that exploits fish buckets NBT to crash servers/realms and it does a pretty good job at it.

We don't need to ban these items to prevent a crash, but instead we replace the items with a new item to delete the NBT.

This effectively stops the crash and allows players to still use the fish buckets in their world.