Much simpler parse handler
Json.stringify() will escape quotes, backlashes, and so on.

The initial goal was just to remove the quotes where needed and subsequently leave quotes where also needed since certain conditions within Minecraft itself handled quotes differently which in turn caused mixed problems with stringify. This approach quickly became a realization that it would be a crappy situation to handle and maintain in the future.

Instead, we use a regex to clean it all up while still using the quotes within the codebase.