Disable modifications to property .nameTag
Modifying the object of type Player property, nameTag, breaks vanilla commands. Or at least it complicates the situation. If Microsoft/Mojang decide to handle this under better terms then this can be reinstated. For now, it has to be disabled so players with permissions in their World can use basic command functions and to prevent command blocks from breaking. The chat ranks will still exist. This patch simply removes the functionality of displaying rank and name above the player's head as seen in Chat.