Updated handler for illegal and salvageable items
Redefined list that is used for hacking but is salvageable so it doesn't ruin gameplay experience. Also check to see if any placed items with a container doesn't contain items because this would clearly be a hack. Only a select few items would contain items within its container during this event.