Bug fix for command deop
It would break if you used the command on yourself so try/catch is used to handle this in such cases. Also handled other messages to be sent out when a condition is already met, otherwise we fall back on a default message. Code is self explanatory so there is no need to use fancy wording here.