Add option to enable/disable shulker boxes
This handles shulker and undyed shulker boxes. Currently it does not appear to be feasible to look inside the containers of these specific blocks at this very moment until the Developers make it possible... Unless everyone is overlooking the issue here with the documentations.

For now, we just give the users control to either allow or not allow shulker boxes. By default shulker boxes are  allowed. You must run the command to enable this feature and not allow shulker boxes, or modify the configuration file if you wish for the setting to persist during server resets as is the normal routine with all configurations for Gametest right now.

This will not ban a player if enabled and a player is found to have a shulker box in their possession. It will simply remove the item from their inventory, or remove it from the world if they manage to attempt the placement of this type of block.