Potential workaround for a bug in minecraft
There seems to be a bug in Minecraft that causes a player entity to briefly return as an entity but not of a player type during the tick. It appears to happen when the angle of the player returns incorrectly. When this happens the script would not recognize the entity as a player and subsequently remove the tag from staff. This patch now attempts to work around the issue by gathering a list of players online filtered out by the tag then verifies if the entity nameTag matches the player as a fail safe to see if the entity is in fact a player or not to prevent accidentally removing the tag due to this unusual bug. [Still testing]