Update illegalitems_c.js
When checking containers we want it to check all items and remove them. So we need it to skip forward to the  next item in the loop. Before, this caused a bug where it would clear the item but move down the script to the other if/else statements and the stored variable would be called to handle the item already removed. In this instance, if item is salvageable but already removed from a nested container then it would unintentionally add that item back to the container and we don't want this behavior.