Add new command !stackban
This command will enable or disable the setting to check players if they have item stacks greater than 64. This is disabled by default. It requires that Illegal Items A, B, or C to be enabled to use this feature. It will let the user know to enable one of those features if it finds that all three are turned off. If this feature is not enabled then it will just clear the illegal stack from the players possession and notify staff about the stack while letting the player know that these stacks are not authorized and no ban will take place.