Update anti jesus
Modified logic for detection. Minecraft allows, whether intentional or not, for players to bunny hop across water. So this feature will naturally flag such players as well. Instead of giving the player a violation we just silently teleport them under water to force them to swim or use another object such as a boat to get across the water because ain't nobody bunny hopping across water in real life anyways.