Update illegalitems_a for illegal enchantments
While checking the player's inventory we quickly verify if each item is enchanted, then we verify what type of enchantment it has and what level that enchantment type is. If the level of that enchantment exceeds or breaks vanilla then we remove the stack from them immediately, warn them that this is not allowed, then pass this information over to staff so they are aware about the player who had the illegally enchanted item.