Fixed random teleport bug.
The issue with the original fly was that if detected only two blocks of air: under the player or in the player. The way player.teleport works is a bit weird, and the ground tag was not checked right after being teleported back - I have no possible explanation why other than a mix of a few bugs. Either way, reverting back to what was used before with a few changes and a reset for the score seems to work even better than before! I still believe in using commands instead of BlockLocation, as it has plenty of bugs, and detects the block it is "most" on, so if you stood on the edge of a block, it would think you were in the air. Jumping on the edge of the said block would increase your fly_timer score, and paired with the ground tag not being checked, you have a false flag.

All of this has been fixed, and note to all others working on fly_a after this: Please don't use BlockLocation unless you can check *at least* the 3 blocks below and on all sides. of the player.