Add list for enchantments
This list is a compilation of what items can currently be enchanted from vanilla gameplay, the types of enchantments those items are allowed to have during vanilla gameplay, and their max levels respectively.

This can be customized for realm/server owners who wish to extend, or restrict which of these items can have which enchantments or how high the levels of those enchantments may be. As is, this is set up for a vanilla experience.