Enhance enchantment detection (illegalitems_c)
This change will not use the data list from enchantmentSlot to verify if the item enchanted is allowed to have that enchantment. If that item is not suppose to have that type of enchantment then it flags the player, notifies staff, warns them, and removes that stack immediately.

Minor correction to illegalitems_b as well.