Major bug fix for illegalitems A, B, and C
This fixes an issue where items were being false flagged and causing them to clear from the player's inventory. It literally comes down to order of operations.

1. Check if item is salvageable
2. Check if item is illegal
3. Check if item stack exceeds allowed amount
4. Check if item has illegal enchantments

Due to how Minecraft functions as a game the order of operations is imperative to prevent or minimize any conflicts with the codebase.