Add new moderation command feature !tester
This command will make a player a tester for the world (server/realm) so they can assist in testing or debugging issues without the repercussions of being hard banned if they trigger Paradox. They will be protected from lockdown and when Paradox does "ban" them for sake of testing then they can rejoin the world again.

To remove them as a tester just run the same command again while they are online and it will clear them.