Update home feature
Minor changes to the code but also included a configuration which allows a limit for how many saved locations can be stored per player. By default this is enabled and the limit is set to 5 per player. It can be changed in config.js by changing the max limit or disabling the module for maximum access that Minecraft can offer.