Update initial script for paradox
Since we subscribe and unsubscribe from these features under certain conditions this causes an issue where if you default the modules to false from config.js but later enable them in-game, the conditional check is not updated from the main script thus causing the feature to not actually turn on. So removing these conditions solves the issues and insures when enabled or disabled that it does what it is instructed to do. I personally find this behavior to be very strange but it's javascript so what can you do.