Fix fly/A
This essentially reverts the code back to commits 787705636196eff8c60dd38579d5024e34eb87eb, ecfb3cc902fe6ef98e3d8b444c36f8b67ac73f3f, b285f759beeb04dae121b196c90c679350987f34, and 3002cf23fb771be80708c6860259bf2ddb218a5c with additional modifications to the code itself to correct some current and past issues.

There was an issue where the player would randomly teleport because the old coordinates were not being updated correctly. This stemmed from when the function for time was executed in the code. This has been move to a more appropriate location for better execution timing. Another issue with the previous code to this change was that it would excessively detect up to 5 blocks below the players Y coordinates to detect air when we only require just the first two. The code also didn't account for the player returning to the ground and would consequently reset the timer for that player. This unfortunately caused a bug that a player could exploit to let them keep flying. In combination to this and the poor execution for the function time this would make the player appear to glitch across the air as it would teleport them back to their old coordinates but their old coordinates would be wrongly updated. Another bug found was players being flagged for flying even if in water. This mostly stemmed from riding in flowing water. Needless to say, this patch accounts for these issues and they should be effectively resolved.