Major fix for allowgmc, allowgma, and allowgms
When you changed the setting to not allow a specific gamemode it unfortunately would set you to that gamemode. This was never intended to be this way. Code is updated to properly check and set players based on the settings.

If Creative Mode is not allowed then it will change the player to Survival. If by chance survival isn't allowed then it will default to adventure. If both are allowed then it will default to Survival.

If Survival Mode is not allowed then it will change the player to Adventure. If by chance adventure isn't allowed then it will default to creative. If both are allowed then it will default to adventure.

If Adventure Mode is not allowed then it will change the player to Survival. If by chance survival isn't allowed then it will default to creative. If both are allowed then it will default to survival.

Big thanks to @Dylanomeneer for bringing this issue to my attention so I could resolve it immediately.