Make sure gamemode settings doesn't lock world
We already check if staff disallow all gamemodes via commands and set it to allow adventure to prevent causing major issues in the world by accident. However, it did not take into account that they could not allow them all from the configuration file and this would consequently still cause issues in the world. This patch will now check to make sure at least one is allowed and if all are set to not allow then default to adventure mode for safety.