Add new utility command for hotbar
This will allow staff to add a hotbar message to display on all players screen. You can configure the message within config.js to persist between server downtime. To use the command here are some basic examples:

Enable with configured message in config.js:
!hotbar

Enable with custom message:
!hotbar message goes here

Turn it off after enabled:
!hotbar

It is very simple and very basic to use.
If you go in Vanish mode while this is enabled then the Vanish message takes precedence over this, and if performance is enabled then that takes precedence over all of them to prevent issues if two or more are turned on in the world.

This module will persist even if there is no message to display. So put a message and share it with others or turn it off.