Add support for Realm Bot Chat Relay
This feature is disabled by default and can be permanently turned on in config.js or temporarily enabled during uptime with the toggle command !rbcr.

This will allow Realm Bot users to send chat messages from within the game to their Discord while using Paradox.

This naturally requires you to actually use Realm Bot and it's chat relay abilities for it to work. Otherwise it does nothing.