Enhance World Border
It is now possible to set your World Border to whatever value you like. No more 1k, 5k, 10k...100k, etc. Just run the command with a value and it will be good to go.

examples:
!worldborder 1234
!worldborder 19867
!worldborder 50000
!worldborder disable
!worldborder help

This is a breaking change! If you have world border enabled prior to this patch then you MUST turn it off before updating. After this is complete then you are safe to enable your world border again.