Correct handling of hotbar
hotbar was using both config and scoreboard and both were being used to determine the status of the utility in terms of it being enabled or disabled. This caused issues as one would stick but the other would reset and give a false impression that the setting was disabled or enabled even though that was not the case. Usage of scoreboard is removed and hotbar specifially targets the configuration file instead now.