Duplicate list of illegal items for JSON
This list is already handled by Gametest but for additional security we are applying this list to the player object as a animation controller.

query.get_equipped_item_name is currently deprecated but not entirely removed yet. This will need to be changed for the current variant query.is_item_name_any.