Update world border to recognize nether and the_end
If you set the world border to 5000, and you move to the nether dimension then it will modify the size of this border by dividing the number by 8. This value can be modified in config.js. By default, the block difference between the overworld and nether is 8 blocks. So 5000 in the overworld would be 625 in the nether. If the player is in the_end then we ignore the world border.

Just to express the size difference of a world border between dimensions.. If you set it to 5000 then you essentially have a overall border size of 10,000 blocks in either direction. When in the nether that 5000 changes to 625 which essentially means you have an overall border size of 1,250 blocks in any direction.

The value for the nether, once again, can be changed in config.js. If you set it to 0, then the code will treat it as a 1 since you can't divide by 0.