Add cooldown to gohome
People are abusing this feature so now there is a cooldown set to 5 minutes by default. If you teleport then you won't be able to use that command for the specified time until that time expires. Staff will be able to change this in config.js. It will accept Days, Hours, Minutes, and Seconds. Can set it to whatever you want. As per usual, staff is excluded from this.