Welcome back .nameTag
This brings back a feature that was removed about 3 months ago in commit 84165a6d41da3cef83a363d4e4ab2375cbb833ed

The commit message in the hyperlink above will explain why it was removed and the issues it caused. In 1.19 this has been fixed and we can safely modify this property to show chat ranks above the players head. A cool feature that I always admired and waited patiently to bring back.

You can disable chat ranks if you do not want this on.