Fix broadcast for staff
The method, SendMsgToPlayer, is not designed to handle certain target selectors. More importantly, the target selectors didn't appear to return a player object to execute the runCommand method from the Player Class and therefore causing the message to not broadcast at all. Since these specific target selectors pertains to all those in the world that meet the specifications then we want to target the method, sendMsg, to effectively filter the target selector and execute the command in a generalized manner.