Add crasher/B
This attempts to catch arrows designed to crash worlds by checking the entitycreate event. If an arrow is created in the world and this event is triggered then it will run through some conditions to see if there are any red flags and kills the arrow if true.

Some ideas pertaining to this commit stem from the works of @MajestikButter and @MrDiamond64. Links provided.

MajestikButter: https://discord.com/channels/523663022053392405/854033525546942464/948349809746669629
MrDiamond64: https://github.com/MrDiamond64/Scythe-AntiCheat/commit/5ae0f70d3a187b6246b9d85a34838162da106d55

In most cases I would provide proper authorship where authorship is due when creating commits but it is not feasible in this instance given the limitations by Git and the mere fact that this has been adopted and adapted to work with Paradox on a different scale. With that said, credit is given where credit is due in the provided links above.