Add tpr command
This will allow non staff members to send a request to another player to teleport to their location. If approved they will be teleported. If not approved then the request will be denied. Only one request can be sent out to each person. If a person is pending a request then your request will be denied autmomatically. There is a 10 second timer with each request that is later verified when the command is executed. If the timer doesn't exist or has expired then your request will be sent. Otherwise your request will be denied.