Update give command to check for NaN
When passing the third parameter to the command it should return a value of some sort to let the command know how much of that item should be given in return. If this parameter is not provided or if it is anything but a number then we catch it and default to a value of 1.