Allow hotbar to be updated efficiently
Originally you would have to disable then enable hotbar if you updated the message. This wasn't really the expected results desired and was left unattended. Now when hotbar is enabled with a message and you want to update the message you can without the hassle of it incidently disabling the feature. The message will update on the screen and it will notify staff that a player updated hotbar.