Fix clearlag
It was not properly clearing items or monsters before. It would seem that by attempting to filter both under the same instance isn't allowed and causes the code to break so each form of entities needed their own functions to be called on. One to handle dropped items and the other to handle monsters.