Update tpr to clear old pendings
In cases where someone sent a request but that request was ignored then it would still be "pending" based on how the code is constructed. This patch should catch those situations and validate that its long expired then clear it. Haven't tested this personally. In theory this code should work out of the box. I'm patching this blind folded lol.