Add whitelist for illegalitems/C
This whitelist will allow illegal items to remain on the list for being illegal but in cases where they are legitimitely registered by Mojang when being placed in the world during the blockPlace Event then we don't want to cause false flags so we ignore such items if found on the whitelist. This whitelist only applies to blocks being placed down. Any other cases and those items will be treated as illegal items and result in a ban.