Cleaning up project environment
~ Executed prettify.
~ Fixed package.json and tsconfig.json to function properly in a linux environment.
~ Corrected npm usage in a local environment in a Linux environment.
~ Insure build directory, node_modules directory, types directory and scripts directory are ignored when pushed upstream.

On a linux environment execute in the terminal from VSC: npm -i

A directory for node_modules will be created and populated by the packages needed for the project. This directory is designated to be ignored by .gitignore.

If you want to build for testing in a BDS environment by quickly accessing the behavior pack in an uncompressed form then execute: npm run build

This will create a build directory at the root of the project folder. All typescripts will be transpiled to javascript and placed within this directory along with all other necessary directories, sub directories, and files required to complete the behavior pack.

To create a distribution pack execute: npm run dist

This will create the build directory in a very similiar fashion as the previous command but then will compress them into a zip ready for shipment found in the build directory.

~ This is just a quick summary on how to use the scripts in package.json under a Linux Environment. The scripts designed to function in a Windows Environment are left intact but not tested in a little while and may require attention.